refactor: do not expose conversion.convertTo
