+++ apis/conversion/src/hypercube.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-prototype-builtins */
 /* eslint-disable guard-for-in */
 /* eslint-disable no-restricted-syntax */
-import createExportFormat from './exportFormat';
+import createExportFormat from './export-format';
 import utils from './utils';
 import arrayUtils from './array-util';
 import helpers from './helpers';
